requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-ollama
faiss-cpu
python-dotenv